Restructuring + basic RV32I structs
